Added another abstraction to the library by introducing a class representing a node connected to the CAN bus.
This node is identified by a textual string name, and may be configured
to be a provider of a set of CAN frame identifiers.

The node abstraction may be useful if you want to maintain a tree of CAN
nodes, and filter CAN bus traffic based on who transmitted the data.